#29 refactor : 추천 영화 목록 표시 기능 개선 - 현재 영화 데이터를 props로 넘겨 장르 검색을 RecommendMovies에서 수행해 동적 장르 검색 기능을 구현
